Revert "Fixed wrong value for showPresentationId in Ticket class"
This reverts commit 21961ed216b58fa9cf625bc778b672c0edf6be0e.

+++ src/main/java/com/ets/gti525/model/Ticket.java
@@ -11,7 +11,7 @@ public class Ticket {
     @GeneratedValue
     private Long id;
 
-    private Long showPresentationId;
+    private int showPresentationId;
 
     public Ticket() {
 
@@ -25,11 +25,11 @@ public class Ticket {
         this.id = id;
     }
 
-    public Long getShowPresentationId() {
+    public int getShowPresentationId() {
         return showPresentationId;
     }
 
-    public void setShowPresentationId(Long showPresentationId) {
+    public void setShowPresentationId(int showPresentationId) {
         this.showPresentationId = showPresentationId;
     }
 }
